rename pull-request command to merge
