+++ src/commands/merge.ts
